added signout and sessions
