+++ package.json
@@ -16,7 +16,7 @@
     "@mui/x-charts": "^7.11.0",
     "@vercel/postgres": "^0.9.0",
     "bcrypt": "^5.1.1",
-    "next": "^14.2.3",
+    "next": "^14.2.5",
     "next-auth": "^4.24.7",
     "react": "^18",
     "react-dom": "^18"
