+++ src/app/page.tsx
@@ -3,15 +3,17 @@ import { Box, createTheme, ThemeProvider } from "@mui/material";
 import Image from "next/image";
 import { FilledButton } from "@/app/components/buttons/filled-button";
 import OutlinedTextField from "@/app/components/textfields/outlined-text-field";
-import { useRouter } from "next/navigation";
+import { redirect, useRouter } from "next/navigation";
 import LockRoundedIcon from '@mui/icons-material/LockRounded';
 import AccountCircle from '@mui/icons-material/AccountCircle';
-import { useState } from "react";
-import { signIn } from "next-auth/react";
+import { useEffect, useState } from "react";
+import { signIn, useSession } from "next-auth/react";
+import { getServerSession } from "next-auth";
+import LogOutIcon from "./components/icons/logout-icon";
 
 const darkTheme = createTheme({
   palette: {
-    mode: "light",
+    mode: "dark",
     primary: {
       main: '#FFA833'
     },
@@ -23,80 +25,30 @@ const darkTheme = createTheme({
   }
 })
 
-export default function Home() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const router = useRouter();
+export default async function Home() {
+  // const router = useRouter();
+  //const session = getServerSession();
+  // const { data: session } = useSession()
 
-    const handleLoginClick = async () => {
-
-      const result = await signIn('credentials', {
-        redirect: false,
-        username,
-        password,
-      });
-
-      if (result?.ok) {
-        router.push('/pages/Dashboard');
-      } else {
-        console.error('Authentication Failed')
-      }
-    };
+  // useEffect(() => {
+  //   const checkSession = () => {
+  //     if (session) {
+  //       router.push('/pages/Dasboard')
+  //     } else {
+  //       router.push('/pages/Login')
+  //     }
+  //   }
+  //   checkSession();
+  // }, [router]);
+  const session = await getServerSession();
   return (
-    <ThemeProvider theme={darkTheme}>
-      <main>
-          <Box sx={{
-              display: 'flex',
-              flexDirection: 'column',
-              minHeight: '100vh',
-              bgcolor: '#000000',
-              justifyContent: 'center',
-              alignItems: 'center',
-          }}>
-              <Box sx={{
-                  display: 'flex',
-                  flexDirection: 'column',
-                  paddingTop: '16px'
-              }}
-              >
-                  <Image
-                  src="/logo.png"
-                  alt="TreatsbyWendoh Logo"
-                  width={359}
-                  height={72}
-                  style={{ marginBottom: '0px' }}
-              />
-              </Box>
-              <Box sx={{
-                  minHeight: '325px',
-                  width: '356px',
-                  display: 'flex',
-                  flexDirection: 'column',
-                  justifyContent: 'center',
-                  alignItems: 'center',
-                  bgcolor: 'transparent',
-                  padding: '2px',
-                  borderRadius: '8px',
-                  
-              }}>
-                  <OutlinedTextField 
-                  placeholder="Username" 
-                  icon={<AccountCircle />}
-                  value={username}
-                  onChange={(e) => setUsername(e.target.value)}/>
-
-                  <OutlinedTextField 
-                  placeholder="Password" 
-                  icon={<LockRoundedIcon />}
-                  value={password}
-                  onChange={(e) => setPassword(e.target.value)}/>
-
-                  <FilledButton onClick={handleLoginClick}>
-                      Login
-                  </FilledButton>
-              </Box>
-          </Box>
-        </main>
-    </ThemeProvider>
-  );
+    <main>
+      {!!session && 
+        redirect('/pages/Dashboard')
+        }
+        {!session && 
+        redirect('/pages/Login')}
+      Loading...
+    </main>
+  )
 }
