+++ src/app/pages/Dashboard/page.tsx
@@ -1,7 +1,10 @@
+import LogOutIcon from "@/app/components/icons/logout-icon";
+
 export default function Page(){
     return(
         <main>
             <h1>
+                <LogOutIcon />
                 Dashboard
             </h1>
         </main>
