--- src/middleware.ts
@@ -1,3 +1,0 @@
-//Without a defined matcher, this one line applies next-auth
-// to the entire project
-export { default } from "next-auth/middleware"
